new exercise

+++ 15-03-2016/terceiro.c
@@ -7,7 +7,7 @@ Crie um programa que leia 2 números e no final informe qual é o maior.
 
 int main() {
   //Declaração das variaveis
-  float a, b;
+  double a, b;
   
   //Imprime nada tela um cabeçalho para o programa
   printf("***************************************\n");
@@ -16,20 +16,20 @@ int main() {
 
   //Socilita e armazena valores paras as variaveis a e b
   printf("Dite o primeiro número: ");
-  scanf("%f", &a );
+  scanf("%lf", &a );
 
   printf("Digite o segundo número: ");
-  scanf("%f", &b);
+  scanf("%lf", &b);
 
   //Verifica se os números digitados são iguais
   if(a == b) { 
-    printf("O número %.2f é igual ao número %.2f\n", a, b);
+    printf("O número %.2lf é igual ao número %.2f\n", a, b);
   }
   //Verifica qual número é maior
   else if (a > b) {
-  	printf("O numero %.2f é maior que %.2f\n", a, b);
+  	printf("O numero %.2lf é maior que %.2f\n", a, b);
   } 
   else {
-  	printf("O numero %.2f é maior que %.2f \n", b, a);
+  	printf("O numero %.2lf é maior que %.2f \n", b, a);
   }
 }
